+++ workers/unity/Assets/Fps/Scripts/SetupLogic/ClientWorkerConnector.cs
@@ -70,7 +70,7 @@ namespace Fps
             if (UseSessionFlow)
             {
                 connectionParams.Network.UseExternalIp = true;
-                builder.SetConnectionFlow(new ChosenDeploymentAlphaLocatorFlow(deployment,
+                builder.SetConnectionFlow(new ChosenDeploymentLocatorFlow(deployment,
                     new SessionConnectionFlowInitializer(new CommandLineConnectionFlowInitializer())));
             }
             else if (Application.isEditor)
@@ -90,10 +90,6 @@ namespace Fps
                     case ConnectionService.Locator:
                         connectionParams.Network.UseExternalIp = true;
                         builder.SetConnectionFlow(new LocatorFlow(initializer));
-                        break;
-                    case ConnectionService.AlphaLocator:
-                        connectionParams.Network.UseExternalIp = true;
-                        builder.SetConnectionFlow(new AlphaLocatorFlow(initializer));
                         break;
                     default:
                         throw new ArgumentOutOfRangeException();